do some shit

+++ Assets/Scripts/Units/Models/Unit/IMovementInfo.cs
@@ -2,6 +2,6 @@ namespace Units.Models.Unit
 {
     public interface IMovementInfo
     {
-        public int MovesLeft { get; set; }
+        public float MovesLeft { get; set; }
     }
 }
