+++ Assets/Scripts/Units/Models/Unit/MovementInfo.cs
@@ -2,6 +2,6 @@ namespace Units.Models.Unit
 {
     public class MovementInfo : IMovementInfo
     {
-        public int MovesLeft { get; set; }
+        public float MovesLeft { get; set; }
     }
 }
